Adding remaining documentation in management client and minor tweaks in authoring client documentation
